1.2

+++ 0x15-file_io/1-create_file.c
@@ -27,7 +27,7 @@ int create_file(const char *filename, char *text_content)
 			len++;
 		}
 		nwrite = write(fd, text_content, len);
-		if (nwrite == -1 || nwrite != len)
+		if (nwrite == -1 ||(size_t)nwrite != len)
 		{
 			close(fd);
 			return (-1);
